refactor(llm-providers): decompose 1.2k-line page into focused components
Splits LlmProviderListPage.vue (was 1212 lines) into:
  - LlmProviderListPage.vue (307 lines) — orchestration + list render
  - components/ProviderCard.vue — card + inline edits + actions
  - components/CreateProviderDialog.vue — Create form + Test gate
  - components/EditCredentialsDialog.vue — Edit + Rotate disclosure + Test re-gate
  - components/TunnelHint.vue — reusable Ollama loopback hint
  - providerHelp.ts — shared provider help / labels map
  - composables/useTestConnectionGate.ts — idle|testing|pass|fail
    state machine with invalidate-on-change watch

Both dialogs now hold a single useTestConnectionGate instance; the
duplicated createTestStatus / editTestStatus state machines are gone.
All data-testids preserved on the same DOM nodes; e2e selectors keep
working. Behaviour is unchanged.
